[IMP] webservice_server_env: move oauth2 UI test and add web dep
