+++ webservice/tests/test_oauth2.py
@@ -8,8 +8,6 @@ from urllib.parse import quote
 import responses
 from oauthlib.oauth2.rfc6749.errors import InvalidGrantError
 
-from odoo.tests import Form
-
 from .common import CommonWebService, mock_cursor
 
 
@@ -217,57 +215,3 @@ class TestWebServiceOauth2WebApplication(CommonWebService):
             json.loads(responses.calls[0].response.content.decode())["access_token"],
         )
         self.assertEqual("cool_token", token["access_token"])
-
-    def test_oauth2_flow_compute_with_ui(self):
-        """Check the ``compute`` method when updating WS from UI"""
-        ws = self.webservice
-        url = self.url
-        form_xmlid = "webservice.webservice_backend_form_view"
-        for auth_type, oauth2_flow in [
-            (tp, fl)
-            for tp in ws._fields["auth_type"].get_values(ws.env)
-            for fl in ws._fields["oauth2_flow"].get_values(ws.env)
-        ]:
-            next_ws_id = ws.sudo().search([], order="id desc", limit=1).id + 1
-            # Create a new WS with each ``auth_type/oauth2_flow`` couple through UI
-            with Form(ws.browse(), form_xmlid) as ws_form:
-                # Common fields
-                ws_form.name = "WebService Test UI"
-                ws_form.tech_name = f"webservice_test_ui_{next_ws_id}"
-                ws_form.protocol = "http"
-                ws_form.url = url
-                ws_form.content_type = "application/xml"
-                ws_form.auth_type = auth_type
-                # Auth type specific fields
-                if auth_type == "api_key":
-                    ws_form.api_key = "Test Api Key"
-                    ws_form.api_key_header = "Test Api Key Header"
-                if auth_type == "oauth2":
-                    ws_form.oauth2_flow = oauth2_flow
-                    ws_form.oauth2_clientid = "Test Client ID"
-                    ws_form.oauth2_client_secret = "Test Client Secret"
-                    ws_form.oauth2_token_url = f"{url}oauth2/token"
-                if auth_type == "user_pwd":
-                    ws_form.username = "Test Username"
-                    ws_form.password = "Test Password"
-            ws = ws_form.save()
-            # Check that ``oauth2_flow`` is the expected one after creation only if the
-            # ``auth_type`` is "oauth2", else it should be False
-            self.assertEqual(
-                ws.oauth2_flow, oauth2_flow if ws.auth_type == "oauth2" else False
-            )
-            # Change WS's ``auth_type`` through UI
-            with Form(ws, form_xmlid) as ws_form:
-                new_auth_type = "none" if ws.auth_type == "oauth2" else "oauth2"
-                ws_form.auth_type = new_auth_type
-                if new_auth_type == "oauth2":
-                    ws_form.oauth2_flow = oauth2_flow
-                    ws_form.oauth2_clientid = "Test Client ID"
-                    ws_form.oauth2_client_secret = "Test Client Secret"
-                    ws_form.oauth2_token_url = f"{url}oauth2/token"
-            ws = ws_form.save()
-            # Check that ``oauth2_flow`` is the expected one after update only if the
-            # ``auth_type`` is "oauth2", else it should be False
-            self.assertEqual(
-                ws.oauth2_flow, oauth2_flow if ws.auth_type == "oauth2" else False
-            )
